+++ CHANGELOG.md
@@ -5,9 +5,9 @@ All notable changes to this project will be documented in this file.
 ## [[NEXT]](https://github.com/iExecBlockchainComputing/iexec-commons-poco/releases/tag/vNEXT) 2023
 
 ### New Features
-- Upgrade `web3j` dependency from 4.8.9 to 4.9.7.
-- Regenerate all wrappers with new web3j cli version
-- Update the readme file with the generation instructions
+- Upgrade `web3j` dependency from 4.8.9 to 4.9.7. (#8)
+- Regenerate all wrappers with new web3j cli version. (#8)
+- Update the readme file with the generation instructions. (#8)
 ### Bug Fixes
 ### Quality
 ### Dependency Upgrades